requirements.txt
streamlit
torch
torchvision
numpy
Pillow
scikit-learn